tutorial level design done
Needs lightning at final parts and lightning fixes.
Next up is the mechanics for the tutorial. !!!
Player starts in the lab. Then via the vent you can go to a new room. In the server room u can get through via the vent. In the vent you will notice that there are guards walking around. 2 guards patrolling. You need to get to the room on the left. Indicate that to the player needs to get there.. etc etc. tutorial finishes at the last room